+++ packages/rights-management-pap-entity-storage/src/index.ts
@@ -1,7 +1,0 @@
-// Copyright 2024 IOTA Stiftung.
-// SPDX-License-Identifier: Apache-2.0.
-export * from "./entities/odrlPolicy";
-export * from "./policyAdministrationPointComponentEntityStorage";
-export * from "./models/IPolicyAdministrationPointComponentEntityStorageOptions";
-export * from "./utils/odrlPolicyConverters";
-export * from "./schema";